requirements.txt
pandas==1.0.1
streamlit==0.62.1
tmdbv3api==1.6.1
requests==2.22.0
beautifulsoup4==4.9.1
Pillow==7.2.0
scikit_learn==0.23.1
